spawn multiple ptys on the same server
Summary:
Prior to this diff, the pty service was a singleton -- it only allowed for a single pty service and if you tried to access that service you would tap into whatever the pty was running. You could not spawn new pty processes.

This diff adds the ability to spawn multiple ptys with a single thrift pty service.

It changes the spawn function to always spawn a new pty, and to return the id of the new pty that was spawned (previously it would just re-use the existing pty). It changes the APIs of other methods to require a pty on which to take action, i.e. `writeInput` now requires the pty id to write input to.

It also changes the internal state of the pty handler to define and maintain an array of pty metadata.

Finally, the "initialCommand" argument was removed from `spawn` since it is better implemented by the client to ensure the pty has started before running the first command.

Reviewed By: stepanhruda

Differential Revision: D13208208

fbshipit-source-id: 8a059188d080eb18270cb5888673955c4de0ebab