Add revisionTree
Summary:
RevisionTree is simply a copy of RevisionStructureTree from interactive smartlog into hg repository client. This type will now be referenced by hg operations, so we needed to hoist it here.

The only difference is that RevisionTree does not contain the optimistic commit logic baked inside of it. This will later be handled how other operations handle optimistic state.

This diff also copies the different PreviewTypes. This type will be completely switched over to the new version in the next diff, since it does not need to depend on the GK.

Reviewed By: aadisriram

Differential Revision: D12913694

fbshipit-source-id: 75ed19931671182ea943fe0ee3cd3fc4fb63e504