Add documentation to operation, use in vcs doc box
Summary:
Add in a command documentation field to HgOperation. This will let us define how to confirm an operation when appropriate. This can also be generally useful information that might be used in other situations, like when hovering on the list of previously run commands. Currently this type is just the same fields as used by VCSCommandDocumentation, but I expect this to diverge over time.

This diff just adds the basic mechanism, in another diff I'll update hide operation to use this for more intelligent documentation.

Reviewed By: aadisriram

Differential Revision: D12987796

fbshipit-source-id: 167096c2bf7b81864580f9f7227a876253e81c0e