Rebase Operation
Summary:
Now that the basics are worked out with Hide, rebase is a real test of how versatile this refactor will be. This diff adds a new operation for rebasing, including preview and optimistic state.

We had a bunch of duplication and weirdness in our hack to add smooth optimistic previews. This consolidated version is much nicer and more accurate! The only thing we lost in the transition is the fullscreen spinner. This can be easily added, but I left it out of this diff.

The only weird thing here is related to drag&drop. Drag&drop requires keeping track of what's being dragged where, and if it's currently dropped or not. This same state was being used to render the preview. This was okay, but now that previews are handled in a more unified way, we want to instead extrapolate an operation out of the drag&drop status. This works just fine, but can be cleaned up a fair bit once we remove the old code path. In the end, the state keeping track of drag&drop status for rebase should only be concerned with drag&drop, and not previewing. This will be nice because we don't need the code for suggested rebases to be mixed with drag&drop. We *could* re-write it now, but that seems unnecessary.

Reviewed By: aadisriram

Differential Revision: D13049287

fbshipit-source-id: 8adf0f339453b13d16177691f3ac1513be58a9fd