Checkout operation
Summary:
Add the operation for check out. It supports hashes/bookmarks/any other revset. Includes optimistic state applier, which was so wonderfully easy to write in this system, compared to what we had before (a mess of conditionals in Revision.js based on multiple parts of state updated in different places)

This should fix a bunch of issues where check out operations would show duplicate/wrong indicators, since it's simpler and tested.

One limitation here that existed before is that we can't show the optimistic destination of an arbitrary revset (since we don't know the hash until hg is done). I don't think it's worth doing an `hg log` call just to show this (it'll be too slow). But we show this correctly for hashes & bookmarks, which seem like 99% of the use case anyway.

Reviewed By: aadisriram

Differential Revision: D13127654

fbshipit-source-id: ca0111757723dfb19ff67ba05d42220c5658ce3f