runOperation
Summary:
Define how to actually run an operation. This is a big function because it will encompass a lot of functionality, but it's removing that functionality from each epic that used to have to deal with all of this. I think with tests this is not too hard to follow.

In this diff:
1. Define a way to observe the current revision tree
2. define a way to observe the global progress of all commands. This is used by smartlog to track current operations and more. This allows centralization of knowing what commands are currently running.
3. define how to run an operation. This includes the following features:
    - check for errors, show atom notifications, detect error messages from stderr on non-zero exit code
    - return progress but also report it to the global progress watcher
    - Ask the operation what the current optimistic state should be, and return that in the progress (also provide a default if the operation does not define this)
    - prevent running multiple commands simultaneously
    - handle kinit and other anomalies

This will eventually also include cancelling operations (it should be easier to do now, since we only need to change this one implementation)

After this refactor, this function will replace almost all of HgRepositoryClient and HgService.

Reviewed By: aadisriram

Differential Revision: D12946908

fbshipit-source-id: eebc9559f5415a7662032e5d71c16802a2099177