Basic HgOperation
Summary:
Add in the basic interface definition for HgOperations and define HgHideOperation as a simple example. This will be extended slightly over time, but I actually don't think much more info will be required other than some documentation stuff. But this should be the place where operations define all of their related information. This includes previews, optimistic state, documentation, etc.

Colocation of all information is what provides a really nice win to this refactor. All of these pieces were strewn across many files and tests before.

This could be potentially generalized into just `Operation`, which could contain a discriminator like `hg | jf | arc | git`, but I see that as being down the line a bit.

See D12987107 for more context and info about how previews and optimistic state actually work here. It was hard to split this up to be defined later in the stack, but logically it belongs there. There are tests for these functions later in the stack too.

Reviewed By: aadisriram

Differential Revision: D12946919

fbshipit-source-id: e9f786e984ab516775d90cdbcaca3b13da9c0964