Always use adb remotely
Summary:
In remote cases we were trying to call adb on the wrong host.

This isn't a permanent solution - we should get rid of passing adbServiceUri completely. For that we will need to use something like adb_proxy so the tunnel for jdwp does not need to be managed by Nuclide.

Reviewed By: Goom11

Differential Revision: D13226041

fbshipit-source-id: 5f2a85308614afdaf7f978a9af7310970455531a