Split scrollbar indicators across multiple canvases, Don't scale...
Summary:
Whenever the cursor moves to a new line, it forces a rerender of the canvas. Computing the visible line number for a file's line number in a way that accounts for folds is somewhat expensive. If you have a large number of diagnostics or a large number of search matches, you don't want to have to rerender/compute these more frequently than you have to.

This allows frequently updating types like cursor to not force the calculation of the positons of these these other, potentially numerous, indicators.

Additionally (sorry to conflate commits!)

Using `image-rendering: pixelated` instead of scale to handle retina displays is way simpler and also actually works (the scale approach was still leading to fuzzy marks for some reason)

Reviewed By: matthewwithanm

Differential Revision: D13162215

fbshipit-source-id: 76b2b6ba1fd3319b3224ae61497b72f2a765a0a0