Replay hot shared observables
Summary:
When subscribing to diagnostic updates, a shared observable subscribed to updates is used to prevent unnecessary work.

This makes it replay the last known value (the set of diagnostics for the system or the file) to new subscribers. Before, since the observable was hot, new subscribers would miss the current state of diagnostics until they changed again.

Reviewed By: mostafaeweda

Differential Revision: D13268030

fbshipit-source-id: 1efc749d62fb92f965204901986db716308c1aa8