Fix BuildOutput Regex stdout msg passthrough
Summary:
`ui.color=always` was adding ANSI codes causing the BUILDOUTPUT message to be filtered out of the processing event stream and causing the debugger to fail due to `no build output`. this diff fixes that bug by adding a special case for messages that (when stripped of ANSI) match the regex query that will be used later to match.

combined: D13268117

Reviewed By: aadsm

Differential Revision: D13250639

fbshipit-source-id: 18f59b815a23008decbc526c8ff36edc7ddf23a6