Fix edit breakpoint condition
Summary:
While implementing the log messages UI I noticed that editing the condition was completely broken and basically not possible to add conditions to breakpoints.

The reason is because `AtomInput` was being used wrongly, it should be either used as a controlled or an uncontrolled component, not a hybrid. This was working fine the way it was just by coincidence. When `AtomInput` was updated to use more recent react APIs here: D13108576, this component stopped working.

I updated this to use `AtomInput` as a controlled component to make it work again.

Reviewed By: ebluestein

Differential Revision: D13293478

fbshipit-source-id: 71534da78f236ed43918d133844a61a6e19131c0