Use regular logging for expressions with no children
Summary:
This is a fix to a bug making some console output to be rendered as regular text instead of teletyped. This is particularly bad in cases where the output spans multiple lines. An example is running lldb commands where everything is shown in one line in variable width text making it almost impossible to read. This is an issue I reported almost 2 weeks ago.

Given other issues around this I debugged it and pinned it down to this diff: D10501045. I read the code related to this change from the debugger to the console and realized this was being caused by all REPL statements now being expressions and sent to the console as such, no matter its origin.

Originally I wanted to make an exception for REPL statements that start with the backtick but this is really implemented at the server level not in Nuclide so no way to detect these unless we assume all statements that start with the backtick should be rendered normally on the console.

I looked into how these were being rendered and realized that expressions coming from an  lldb command or simple expressions like `2+2` were being rendered by `SimpleValueComponent`. In this component they are being rendered by the `renderDefault` function that just outputs the value without any styling and formatted like `<name>: <expression value>`. The name being there is also redundant as that already exists in the previous line of the console in the debugger case.

To avoid this my solution is to only render as an object expressions that have children (these are rendered as `ExpressionTreeComponent` like struct's, strings, etc.). The other ones like `2+2` and lldb commands do not have children and can be output as regular text console.

If there's a better solution for this (as I might be lacking context), I'm happy to implement it.

Before:
{F145337723}

After:
{F145337724}

PS: Unrelated to this diff but related to the breaking diff: Expressions contain the configuration of the debugger which is not serializable, this breaks storing the state of the nuclide console. You can see that by the exceptions shown on the chrome console.

Reviewed By: ebluestein

Differential Revision: D13290530

fbshipit-source-id: 0b7281639e544143fd7afbdd32c2fb188bfdd4d3