Gate the Log Message UI in Edit Breakpoints behind a GK
Summary:
I added a GK (`nuclide_debugger_logging_breakpoints`) to gate the UI of this new feature as it can only be released once the new lldb-vscode is published.

I'm not sure if this is the best way to check the GK though. The reason why I have the `_supportsLoggingBreakpoints` function is because I'm expecting us to add a capability to indicate this support just like we have for the conditions, but for now it only checks the GK.

Reviewed By: ebluestein

Differential Revision: D13293481

fbshipit-source-id: a19c66c80e2dea5415ad402357e841ebd8c18a84