Add Log Message UI to the Edit Breakpoint panel
Summary:
Add the a Log Message field to the "Edit Breakpoint" panel. Nothing special, it's basically a copy of the condition with updated text. I'm still using the same css classes as they seem appropriate as well and didn't want to clone.

In this diff this new field is not gated, I did this in the next one to avoid creating a big diff with 2 different concerns.

**Concerns:**
* Not all debuggers support this feature. Should I just add a message saying that? Not sure what is the best way to deal with that. I saw that we check a capability for conditions but we default to true if we don't have a running debugger to check that.
* In the UI it doesn't seem that there's a clear separation between the condition and the log message, specially when both are filled in. Not sure what would be the best approach for this or if this how we're doing everywhere else.

{F145400804}

Reviewed By: ebluestein

Differential Revision: D13293480

fbshipit-source-id: df6fb3fd7ca71809f1329b4a54160f3ed2bd7b24