Enable working revision selection in uncommitted changes dropdown
Summary:
In the file tree's uncommited changes section, you can now change what revision the split view will compare to using that section's dropdown.

This was done by adding a list of head to fork base revisions to the file tree's state.

This list will only take into account revisions for the current working directory: any other project/repo will default to head revision. This list is updated by subscribing to cwdApi, then based on the current working directory, finding its backing repo and subscribing to its revision changes.

This list will be used to populate the dropdown.

Note that in the current state, the diff view navigator's history UI and the dropdown are not synced: if you change the current working revision in the file tree, it will not be reflected in the diff view navigator.

Reviewed By: tjfryan

Differential Revision: D13248475

fbshipit-source-id: e729ec1d7608fdce1e13036233d39b765eb71882