Add config option for merge3
Summary:
We don't respect user's hg config `ui.merge=:merge3` setting becuase `ui.merge` can be used to open interactive editors that break nuclide. This means users who set it to `merge3` won't get the merge3 functionality.

This diff adds a config option to the smartlog package to allow merge3.

This is actually really annoying to thread through, since each of our hg calls has its own config and we don't have a good way to pass an option from the client (the others are all determined on the server, which won't work for a config setting).

So, I decided this change will ONLY work with the new Hg Operations refactor. Since that will be launched everywhere soon enough, it should be ok. Although we could consider delaying landing this until it's ready to go. (at least rebase maybe?)

Additionally, I had to add a setter to hgRepositoyClient which feels a little wrong. Another solution would be to pass this to every call of `repository.runOperation()`, but that's also ugly.

Also worth noting, this will not work with `arc pull` merge conflicts (because of course not :'( ).

Reviewed By: aadisriram

Differential Revision: D13329874

fbshipit-source-id: a55fb11717fb9f8ff2ea43c2307af152993477fd