Add Log Message to the Breakpoints View
Summary:
Added a new row to the "Breakpoints View" so we can see the log message as well as the condition.
Also gated based on the GK, again, not sure if this is the best way to gate it.
I reused the same css classes because they were just the same as the condition.

PS: Should I call this "Logging" instead?

{F145400990}

Reviewed By: ebluestein

Differential Revision: D13293482

fbshipit-source-id: e550fe199f9bade53df2d8efa81eec07dc9e250b