Add showFullscreenSpinner
Summary: Add a full screen spinner to HgPullOperation, so that it works like it used to. This is done by adding a field to the `ReportedOptimisticState` from the previous diff. This field is treated slightly differently than the `optimisticApplier`, in that it is only cleared when the operation returns null optimistic state or the command has completely finished. This way, pull can show a spinner without needing to wait for a refresh (simply `Observable.of({showFullscreeSpinner: true})`)

Reviewed By: aadisriram

Differential Revision: D13249058

fbshipit-source-id: edac51bedfeb6859b2a0e4bfa571909635932676