Return ReportedOptimisticState from HgOperations
Summary:
Instead of having each operation retun the optimisticApplier function directly, wrap it in an object so that in the future we can pass additional optimistic/progress data in the same function. This will be convenient, because all optimistic/progress state is typically known at the same time.

This object can be extended with fields such as:
- `showFullscreenSpinner` so hg pull works like it used to (rebase too)
- `optimisticSuccessions` could hold `oldHash` -> `newRevisionInfo` so that the commit info view can show the latest optimistic data in an easy way
- some sort of per-commit inline progress bar, similar to jf submit. This could be useful if we port jf submit to the same system.

Reviewed By: aadisriram

Differential Revision: D13249059

fbshipit-source-id: c431590ec12cf6497c9142744824719304105ac3