Fix "react-dom/profiling" redbox
Summary:
We've been rewriting all "react-dom" require calls to "react-dom/profiling", regardless of the origin of the request. However, third-party packages that use react-dom are unlikely to have "react-dom/profiling" in their `node_modules`. This would cause a redbox.

With this change, we only rewrite the require calls coming from Nuclide's source tree.

Reviewed By: captbaritone

Differential Revision: D13333775

fbshipit-source-id: 4069ee68a804fa2c93f337329da71603d02c0bc1